requirements.txt
matplotlib
numpy
tqdm
pyyaml
opencv-python
wandb
ipykernel